Правки к дз
