+++ js/api.js
@@ -1,13 +1,9 @@
-import {SLICE} from './data.js';
-import {createSimilarPopups} from './map.js';
-import {onSuccess} from './on-event.js';
-import {showAlert} from './util.js';
+import {showAlert} from './alert.js';
 
-export const getData = function () {
+export const getData = function (onSuccess) {
   return fetch('https://22.javascript.pages.academy/keksobooking/data')
-    .then(onSuccess)
     .then(response => response.json())
-    .then(data => createSimilarPopups(data.slice(SLICE.min, SLICE.max)))
+    .then(data => onSuccess(data))
     .catch(() => {
       showAlert('Не удалось получить данные. Попробуйте обновить страницу');
     });
