+++ js/script.js
@@ -1,6 +1,3 @@
-import {getData} from './api.js';
 import {createMap} from './map.js';
 
 createMap();
-
-getData();
